Change CopyConfirmHandler to pre-allocate byte buffer and reallocate it only if size is not enough.
Testing done
Reader java writer cpp
Writers=1 Readers=1 BufSizeKB=65536 BurstSize=2000000 MessageSizeB=1 BackpressureSleepMs=50 FLAVOR=FIRST_LAST_DURATION Iterations=10

Before
Avg-throughput-first-last: 7944.673610761601 (Kmsg/sec)

After
Avg-throughput-first-last: 9155.797895283491 (Kmsg/sec)

+++ java/toroni/src/main/java/toroni/rmp/CopyConfirmCallback.java
@@ -5,8 +5,6 @@
 
 package toroni.rmp;
 
-import com.sun.jna.Pointer;
-
 /**
  * Copies data to local storage and when confirmed invokes a callback.
  */
@@ -14,16 +12,15 @@ public interface CopyConfirmCallback {
 
   /**
    * Copies {@code length} bytes starting from {@index} data to local storage.
-   * 
+   *
    * @param index
    * @param length
    * @return true if successful, false otherwise
    */
-  boolean copy(long index, long length);
+  boolean copy(long index, int length);
 
   /**
    * Invokes a callback once the data has been confirmed.
    */
   void confirm();
-
 }
